requirements.txt
numpy
matplotlib
scipy
pyyaml
numba
ducc0
simms
codex-africanus
matplotlib
SciencePlots
tqdm
zarr
jupyter
ipywidgets
https://github.com/ratt-ru/QuartiCal/archive/refs/heads/v0.2.0-dev.zip
https://github.com/ratt-ru/pfb-clean/archive/refs/heads/master.zip
